Update instructions for using scenario 3 for multiple properties
I'm setting up the ga360 block with a bigquery connection and have run into a problem I'm not sure how to solve.
This is the error message I'm getting: Failed to retrieve data - Name partition_date not found inside ga_sessions at [9:24]
this is where partition_date shows up in the query: "WHERE 
	(((ga_sessions.partition_date) >= ((TIMESTAMP_ADD(TIMESTAMP_TRUNC(CURRENT_TIMESTAMP(), DAY), INTERVAL -7 DAY))) AND (ga_sessions.partition_date) < ((TIMESTAMP_ADD(TIMESTAMP_ADD(TIMESTAMP_TRUNC(CURRENT_TIMESTAMP(), DAY), INTERVAL -7 DAY), INTERVAL 7 DAY)))))"
I'm using scenario #3 to handle multiple properties. And commented out the sql clause for the partition_date field as instructed in the block comments.

Solution was to change from `dimension:` to `filter:` on the `partition_date` field so it's only included in the sql_table_name view condition for the FROM clause, but not be in the WHERE clause. I removed the note about changing the field `type:` because it's already that `type:`.